requirements.txt
pytest
requests
pyyaml
allure-pytest